hkdf: increase MAX_HKDF_INFO_LEN
Encrypted client hello (ECH) offer confirmation requires computing an 8 byte
confirmation value:

```
   accept_confirmation = HKDF-Expand-Label(
      HKDF-Extract(0, ClientHelloInner.random),
      "ech accept confirmation",
      transcript_ech_conf,
      8)
```

Similarly, in the hello-retry request confirmation case we compute:
```
   hrr_accept_confirmation = HKDF-Expand-Label(
      HKDF-Extract(0, ClientHelloInner1.random),
      "hrr ech accept confirmation",
      transcript_hrr_ech_conf,
      8)
```

The HKDF-Expand-Label and HKDF-Extract algorithms are unmodified from RFC 8446.

For a handshake using SHA-384, or SHA-512, as the digest algorithm it's
possible for the `info` parameter to the HKDF expand step to exceed 80
bytes.

```
  2 bytes for the output len
  1 byte for the label len
  6 bytes for the label prefix
 23 bytes for the ECH confirmation label (or 27 for HRR)
  1 byte for the context len
 48 bytes for the context (SHA-384), or 64 bytes (SHA-512)
------------------------------------------
= 81 or 97 bytes total, SHA-384, non-hrr
= 85 or 101 bytes total, SHA-512, hrr

Both would exceed the existing `MAX_HKDF_INFO_LEN` limit of 80 bytes,
and so produce `Unspecified` errors.

This commit increase the limit to 102 bytes (it seemed a nicer value
than 101), allowing the use of aws-lc-rs HKDF for ECH confirmation
purposes without panic.