cipher: clarify the EncryptingKey docs (#488)
Previously the `aws_lc_rs::cipher::EncryptingKey` type's `encrypt()` and
`less_safe_encrypt()` functions were documented as returning a reference
to the encrypted data. However, the implementation uses an `in_out: &mut
[u8]` argument for the encrypted data and **returns**
a `DecryptionContext`, not a ref. to ciphertext.

Similarly, it feels important to call out that the `DecryptionContext`
when using `encrypt()` contains a randomly generated IV. This makes
usage clearer and also provides good counter-point to the disclaimer on
`less_safe_encrypt()` about the provenance (and safety) of the IV being
unknown.

Co-authored-by: Justin W Smith <103147162+justsmth@users.noreply.github.com>