aws-lc-fips-sys: Default to clang compiler
The underlying aws-lc library is broken when built with FIPS support on GCC-14,
which comes by default with many Linux distributions.

When this crate is used as a sub-dependency, the users' only workaround is to force
all C/C++ code on other crates to also be compiled with clang by setting the CC/CXX environment variables.

This change circumvents the issue by setting the CC/CXX environment variables in the crate
build script.

This is a temporary workaround until the underlying issue is resolved:
https://github.com/aws/aws-lc/issues/2010

Signed-off-by: Michel Heily <michel.heily@wiz.io>