Drop explicit/pinned libloading dependency
Unlike clap and criterion, which are pinned dev-dependencies, libloading
shows up in downstream dependency graphs. As such, pinning libloading
to 0.8.8 in these crates holds back its use in all downstream projects,
the vast majority of which will have a newer MSRV. As such, pinning is
not a great solution for the problem at hand. Instead, adjust CI workflows
to make sure an older libloading is used.