Respect LDFLAGS in memcmp_check (#958)
This check invokes the compiled executable and hence needs to link it.
CMake handles this via LDFLAGS but `cc` doesn't. In setups with custom
linker setups this could lead to a mismatch between the expected and the
actually used linker. Explicitly respecting LDFLAGS here brings us back
to parity with CMake.

Fixes https://github.com/aws/aws-lc-rs/issues/858 whose resolution is
not applicable if ignoring compiler warnings is not an option.

Co-authored-by: Aaron Siddhartha Mondal <aaron.mondal@gmail.com>